backend/requirements.txt
# altair==5.0.1
# annotated-types==0.5.0
# anyio==3.7.1
# async-timeout==4.0.3
# asyncpg==0.28.0
# attrs==24.2.0
# backports.zoneinfo==0.2.1
# blinker==1.6.3
# cachetools==5.5.2
# certifi==2026.1.4
# charset-normalizer==3.4.4
# click==8.1.8
# colorama==0.4.6
# databases==0.8.0
# decorator==5.1.1
# exceptiongroup==1.3.1
# fastapi==0.103.2
# gitdb==4.0.12
# gitpython==3.1.46
# greenlet==3.1.1
# h11==0.14.0
# idna==3.10
# importlib-metadata==6.7.0
# importlib-resources==5.12.0
# jinja2==3.1.6
# jsonschema==4.17.3
# markdown-it-py==2.2.0
# MarkupSafe==2.1.5
# mdurl==0.1.2
# numpy==1.21.6
# packaging==23.2
# pandas==1.3.5
# Pillow==9.5.0
# pkgutil-resolve-name==1.3.10
# protobuf==4.24.4
# psycopg2-binary==2.9.9
# pyarrow==12.0.1
# pydantic==2.5.3
# pydantic-core==2.14.6
# pydeck==0.8.1b1
# pygments==2.17.2
# Pympler==1.1
# pyrsistent==0.19.3
# python-dateutil==2.9.0.post0
# python-dotenv==0.21.1
# pytz==2025.2
# pytz-deprecation-shim==0.1.0.post0
# pywin32==308
# redis==5.0.8
# requests==2.31.0
# rich==13.8.1
# six==1.17.0
# smmap==5.0.2
# sniffio==1.3.1
# SQLAlchemy==1.4.54
# starlette==0.27.0
# streamlit==1.23.1
# tenacity==8.2.3
# toml==0.10.2
# toolz==0.12.1
# tornado==6.2
# typing-extensions==4.7.1
# tzdata==2025.3
# tzlocal==4.3.1
# urllib3==1.26.20
# uvicorn==0.22.0
# validators==0.20.0
# watchdog==3.0.0
# zipp==3.15.0

fastapi
uvicorn
groq
python-dotenv
sqlalchemy